% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{calculate_cp_funnel}
\alias{calculate_cp_funnel}
\title{Calculate content progress funnel.}
\usage{
calculate_cp_funnel(sc = state_changes)
}
\arguments{
\item{sc}{A data table: (user_id, tree_id, date, state, statediff).
The result of calling cctimes::calculate_state_changes.}
}
\value{
A data table: (user_id, tree_id, start_date, conversion_date,
completion_date).
}
\description{
Calculate content progress funnel.
}

