% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{calculate_cp_times}
\alias{calculate_cp_times}
\title{Calculate content progress times.}
\usage{
calculate_cp_times(sc = state_changes)
}
\arguments{
\item{sc}{A data table: (user_id, tree_id, date, state, statediff).
The result of calling cctimes::calculate_state_changes.}
}
\value{
A data table: (user_id, tree_id, start_date, conversion_date,
completion_date, start_to_conversion, conversion_to_completion).
}
\description{
Calculate content progress times.
}

