% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{calculate_made_it_score}
\alias{calculate_made_it_score}
\title{Calculate 'made it' score.}
\usage{
calculate_made_it_score(cpt = cp_times)
}
\arguments{
\item{cpt}{A data table - the result of calling cctimes::calculate_cp_times}
}
\value{
A data table (user_id, tree_id, date_type,
content_progress_date, made_it) that tells us if the appropriate type of
content progress was achieved within 30 days.
}
\description{
Calculate 'made it' score.
}

