% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{calculate_state_changes}
\alias{calculate_state_changes}
\title{Identify events when content changes state.}
\usage{
calculate_state_changes(ccc = content_completion_clean)
}
\arguments{
\item{ccc}{A data frame of content completion times. The result of calling
cctimes::delete_single_pagers.}
}
\value{
A data table: (user_id, tree_id, date, state, statediff).
}
\description{
Identify events when content changes state.
}

