% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weekly_data.r
\name{calculate_weekly_progress_rates}
\alias{calculate_weekly_progress_rates}
\title{Calculate weekly progress rates.}
\usage{
calculate_weekly_progress_rates(mis = made_it_scores, week_sequence)
}
\arguments{
\item{mis}{A data table: the result of calling calculate_made_it_score}

\item{week_sequence}{A data table: (week_beginning) consisting of dates
spaced one week apart. The result of calling glootility::create_week_sequence.
tree_id, date_type, made_it)}
}
\value{
A data table (week_beginning, date_type, pct_made_it) that indicates
the percentage of program starts that were converted to 'in progress' and
the percentage of 'in progress' that were finished within 30 days.
}
\description{
Calculate weekly progress rates.
}

