% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.r
\name{create_week_list}
\alias{create_week_list}
\title{Create week list given a start and end date.}
\usage{
create_week_list(start_date = as.Date("2016-01-01"), end_date = Sys.Date(),
  week_begins = 2)
}
\arguments{
\item{start_date}{The earliest date allowed in the output.}

\item{end_date}{The latest date allowed in the output.}

\item{week_begins}{The day of the week that all output dates must fall on.
Given as an integer (1 = Sunday , 2 = Monday , etc...).}
}
\value{
A data table (week_beginning) of dates that correspond to the first
date in each week.
}
\description{
Create week list given a start and end date.
}

