% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{delete_single_pagers}
\alias{delete_single_pagers}
\title{Delete single-page content.}
\usage{
delete_single_pagers(tn = treeid_nodecount, cc = content_completion)
}
\arguments{
\item{tn}{A data frame: (tree_id, node_count).}

\item{cc}{A data frame  of content completion/conversion times.}
}
\value{
A data table of content completion/conversion times, minus all
trees with a single page only.
}
\description{
Delete single-page content.
}

