% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.r
\name{pull_data}
\alias{pull_data}
\title{Pull data from Looker and Postgres}
\usage{
pull_data(looks = cctimes::look_list, queries = cctimes::query_list,
  con = postgres_connection$con)
}
\arguments{
\item{looks}{Looks pointing to Looker data neccessary for this project.}

\item{queries}{Queries to get Postgres data}
}
\value{
A list of data frames, each the result of a single look or query.
}
\description{
Pull data from Looker and Postgres
}

